pr3 pre-release check
